Redesign Linguist with Stitch design system principles
- Introduced DESIGN.md to establish a formal design system.
- Refactored AppComponent into a layout shell with routing.
- Moved search logic and UI to HomeComponent.
- Modernized UI with consistent CSS variables, Pridi/Roboto typography, and card-based results.
- Updated Angular style budgets and fixed unit tests.
