+++ src/app/app.component.css
@@ -1,115 +1,1 @@
-.container {
-  max-width: 800px;
-  margin: 2rem auto;
-  padding: 0 1rem;
-  font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-}
-
-.search-section {
-  display: flex;
-  gap: 0.5rem;
-  margin-bottom: 2rem;
-}
-
-.search-input {
-  flex: 1;
-  padding: 0.75rem;
-  font-size: 1rem;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-}
-
-.search-button {
-  padding: 0.75rem 1.5rem;
-  font-size: 1rem;
-  background-color: #007bff;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-  transition: background-color 0.2s;
-}
-
-.search-button:hover {
-  background-color: #0056b3;
-}
-
-.loading {
-  text-align: center;
-  margin: 1rem 0;
-  font-style: italic;
-  color: #666;
-}
-
-.error {
-  color: #dc3545;
-  margin: 1rem 0;
-  padding: 0.75rem;
-  background-color: #f8d7da;
-  border: 1px solid #f5c6cb;
-  border-radius: 4px;
-}
-
-.history-sequence {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  gap: 0.5rem;
-  margin-bottom: 2rem;
-  padding: 1rem;
-  background-color: #f8f9fa;
-  border-radius: 4px;
-}
-
-.history-item {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.history-word {
-  cursor: pointer;
-  color: #007bff;
-  font-weight: bold;
-}
-
-.history-word:hover {
-  text-decoration: underline;
-}
-
-.separator {
-  color: #6c757d;
-}
-
-.etymology-display {
-  background-color: white;
-  padding: 1.5rem;
-  border-radius: 8px;
-  box-shadow: 0 2px 4px rgba(0,0,0,0.1);
-}
-
-.etymology-display h2 {
-  margin-top: 0;
-  border-bottom: 2px solid #eee;
-  padding-bottom: 0.5rem;
-  text-transform: capitalize;
-}
-
-.etymology-text {
-  line-height: 1.6;
-  font-size: 1.1rem;
-  color: #333;
-}
-
-.clickable-word {
-  cursor: pointer;
-  color: #007bff;
-  transition: color 0.1s;
-}
-
-.clickable-word:hover {
-  color: #0056b3;
-  text-decoration: underline;
-  background-color: #e9ecef;
-  border-radius: 2px;
-}
+/* AppComponent specific styles (layout is mostly handled by global .main and HeaderComponent) */
