+++ src/styles.css
@@ -1,37 +1,53 @@
-/* You can add global styles to this file, and also import other style files */
+/* Global styles and CSS variables based on DESIGN.md */
 @import "normalize.css";
+
+:root {
+  --primary-color: #1a237e;
+  --accent-color: #b8860b;
+  --bg-color: #fdfdfb;
+  --surface-color: #ffffff;
+  --text-primary: #2c3e50;
+  --text-secondary: #546e7a;
+  --border-color: #e0e0e0;
+  --error-color: #c62828;
+  --success-color: #2e7d32;
+  --max-width: 900px;
+  --shadow: 0 2px 10px rgba(0,0,0,0.05);
+  --border-radius: 8px;
+}
 
 html,
 body {
-    margin: 0;
-    font-family: Roboto, "Helvetica Neue", sans-serif;
-    line-height: 1 !important;
+  margin: 0;
+  font-family: 'Roboto', sans-serif;
+  background-color: var(--bg-color);
+  color: var(--text-primary);
+  line-height: 1.6;
+}
+
+h1, h2, h3, h4, h5, h6 {
+  font-family: 'Pridi', serif;
+  color: var(--primary-color);
+  margin-top: 1.5rem;
+  margin-bottom: 1rem;
 }
 
 .main {
-    max-width: 700px;
-    margin: 0 auto;
+  max-width: var(--max-width);
+  margin: 0 auto;
+  padding: 0 20px;
 }
 
-.content {
-    display: flex;
-    justify-content: space-around;
-    width: 100%;
-    max-width: 700px;
-    margin-bottom: 3rem;
+a {
+  color: var(--primary-color);
+  text-decoration: none;
+  transition: color 0.2s;
 }
 
-.content h1 {
-    margin-top: 1.75rem;
+a:hover {
+  color: var(--accent-color);
 }
 
-.content p {
-    margin-top: 1.5rem;
-}
-
-@media screen and (max-width: 650px) {
-    .content {
-        flex-direction: column;
-        width: max-content;
-    }
+* {
+  box-sizing: border-box;
 }
